requirements.txt
arena-py
openai